chore: refactor main to home
